An ugly trick to implement conconcurrent invalidations
SSSP master co-routine spawns multiple tasks, each of which spawns another
multiple tasks to send concurrent invalidations. These concurrent tasks must be
executed before a writer finishes. But is is possible that all co-routines are
executing writers and readers which wait for those writers, which leads to a
deadlock like priority inversion. The solution is that dedicated co-routines are
necessary to execute invalidation tasks only. This approach ensures the progress
of invlidations and writers.